Move acquisition details components to their own folder
