+++ source/frontend/src/features/properties/map/acquisition/detail/fileDetails/AcquisitionOwnersSummaryView.tsx
