+++ source/frontend/src/features/properties/map/acquisition/detail/fileDetails/AcquisitionSummaryView.test.tsx
